+++ src/repo/uow.ts
@@ -1,4 +1,0 @@
-export interface UnitOfWork {
-	transactional<T>(fn: () => Promise<T>): Promise<T>;
-}
-export type RepoProvider<R> = (uow: UnitOfWork) => R;